Recommiting lightcurve changes.
